+++ xavier/core/allocator.h
@@ -9,20 +9,20 @@ namespace xv::core
     struct Allocator
     {
     protected:
-        uint64_t allocated = 0;
+        usize allocated = 0;
 
     public:
         Allocator() = default;
         Allocator(const Allocator &) = delete;
         Allocator &operator=(const Allocator &) = delete;
-        virtual std::shared_ptr<Buffer> alloc(uint64_t nbytes) = 0;
+        virtual std::shared_ptr<Buffer> alloc(usize nbytes) = 0;
         virtual void free(std::shared_ptr<Buffer> buff) = 0;
-        uint64_t get_allocated() const { return allocated; }
+        usize get_allocated() const { return allocated; }
     };
 
     struct CommonAllocator : public Allocator
     {
-        std::shared_ptr<Buffer> alloc(uint64_t nbytes) override
+        std::shared_ptr<Buffer> alloc(usize nbytes) override
         {
             allocated += nbytes;
             auto ptr = new uint8_t[nbytes];
